refactor(sikte): split into modules
