+++ sikte/src/perf_events/mod.rs
@@ -1,2 +1,1 @@
 pub mod main;
-mod table;
